Revert "Support UIScene"
