+++ Cartfile.resolved
@@ -1,2 +1,6 @@
 github "chronicqazxc/WHCoreServices" "1e190c062d927eedda0834054f9f4866422aa46a"
 github "chronicqazxc/WHPromise" "b85d26b1832eef9e585c6ae90ae0d67092c41329"
+github "facebook/FBAllocationTracker" "0.1"
+github "facebook/FBMemoryProfiler" "0.1.3"
+github "facebook/FBRetainCycleDetector" "0.1"
+github "krzysztofzablocki/LifetimeTracker" "1.7.0"
